[dev] feat (index.test): scraping & insert
