+++ test/index.test.js
@@ -1,0 +1,8 @@
+
+const { main } = require('../index');
+
+async function test () {
+  console.log(await main);
+}
+
+test();
