+++ test/services/scraping/index.test.js
@@ -1,0 +1,8 @@
+const { scrape } = require('../../../services/scraping');
+
+async function main () {
+  const data = await scrape();
+  console.log(data)
+  return data;
+}
+main();
